requirements.txt
Django==4.1.2
django-admin-sortable2==2.1.3
environs==9.5.0
Pillow==9.2.0
